restoring changes

+++ src/main/java/com/backend/Retrospect/room/repository/IRoomRepository.java
@@ -4,7 +4,7 @@ package com.backend.Retrospect.room.repository;
 import com.backend.Retrospect.room.entity.RoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRoomRepository extends JpaRepository<RoomEntity,Integer>
+public interface IRoomRepository extends JpaRepository<RoomEntity,Long>
 {
 
 }
